requirements.txt
kafka-python==2.0.2
pyspark==3.5.0
elasticsearch==8.11.0
requests==2.31.0